pyproject.toml
[project]
name = "pipeline-api-bitcoin-streamlit"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexandre Pinheiro",email = "alexandre.ypinheiro@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.1"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "streamlit (>=1.51.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
